+++ packages/typespec-azure-core/src/rules/require-docs.ts
@@ -1,4 +1,18 @@
-import { Enum, Model, Operation, createRule, getDoc, paramMessage } from "@typespec/compiler";
+import {
+  Discriminator,
+  Enum,
+  Model,
+  ModelProperty,
+  Operation,
+  Program,
+  Union,
+  createRule,
+  getDiscriminatedTypes,
+  getDiscriminator,
+  getDoc,
+  paramMessage,
+} from "@typespec/compiler";
+import { getVersionsForEnum } from "@typespec/versioning";
 import {
   isExcludedCoreType,
   isInlineModel,
@@ -6,6 +20,48 @@ import {
   isTemplatedInterfaceOperation,
   isTemplatedOperationSignature,
 } from "./utils.js";
+
+/** Versioning enums are self-documenting and don't need separate documentation. */
+function isExcludedVersionEnum(program: Program, enumObj: Enum): boolean {
+  const versions = getVersionsForEnum(program, enumObj);
+  if (versions !== undefined && versions.length > 0) {
+    return true;
+  }
+  return false;
+}
+
+function findDiscriminator(program: Program, model?: Model): Discriminator | undefined {
+  if (!model) return undefined;
+  const disc = getDiscriminator(program, model);
+  if (disc) {
+    return disc;
+  }
+  return findDiscriminator(program, model.baseModel);
+}
+
+/** Discriminator enums and unions are self-documenting and don't need separate documentation. */
+function isExcludedDiscriminator(
+  program: Program,
+  type: ModelProperty | Enum,
+  discTypes: [Model | Union, Discriminator][]
+): boolean {
+  if (type.kind === "ModelProperty") {
+    const disc = findDiscriminator(program, type.model);
+    if (disc && disc.propertyName === type.name) {
+      return true;
+    }
+  } else if (type.kind === "Enum") {
+    for (const [discType, discName] of discTypes) {
+      if (discType.kind === "Model") {
+        const discObj = discType.properties.get(discName.propertyName);
+        if (discObj?.type === type) {
+          return true;
+        }
+      }
+    }
+  }
+  return false;
+}
 
 export const requireDocumentation = createRule({
   name: "documentation-required",
@@ -15,8 +71,16 @@ export const requireDocumentation = createRule({
     default: paramMessage`The ${"kind"} named '${"name"}' should have a documentation or description, use doc comment /** */ to provide it.`,
   },
   create(context) {
+    const discTypes = getDiscriminatedTypes(context.program);
     return {
       enum: (enumObj: Enum) => {
+        // version enums and enum members are considered intrinsically self-documenting
+        if (isExcludedVersionEnum(context.program, enumObj)) {
+          return;
+        }
+        if (isExcludedDiscriminator(context.program, enumObj, discTypes)) {
+          return;
+        }
         if (!getDoc(context.program, enumObj) && !isExcludedCoreType(context.program, enumObj)) {
           context.reportDiagnostic({
             target: enumObj,
@@ -72,6 +136,10 @@ export const requireDocumentation = createRule({
             });
           }
           for (const prop of model.properties.values()) {
+            // Properties that are discriminators are considered self-documenting
+            if (isExcludedDiscriminator(context.program, prop, discTypes)) {
+              return;
+            }
             if (!getDoc(context.program, prop)) {
               context.reportDiagnostic({
                 target: prop,