+++ packages/typespec-client-generator-core/src/internal-utils.ts
@@ -5,6 +5,7 @@ import {
   Namespace,
   Operation,
   Program,
+  ProjectedProgram,
   Type,
   Union,
   createDiagnosticCollector,
@@ -19,6 +20,7 @@ import { HttpOperation, HttpStatusCodeRange } from "@typespec/http";
 import { getAddedOnVersions, getRemovedOnVersions, getVersions } from "@typespec/versioning";
 import {
   SdkBuiltInKinds,
+  SdkClient,
   SdkEnumType,
   SdkHttpResponse,
   SdkModelPropertyType,
@@ -105,7 +107,7 @@ export function updateWithApiVersionInformation(
   return {
     isApiVersionParam,
     clientDefaultValue: isApiVersionParam ? context.__api_version_client_default_value : undefined,
-    onClient: isApiVersionParam,
+    onClient: onClient(context, type),
   };
 }
 
@@ -128,7 +130,8 @@ export function getAvailableApiVersions(context: TCGCContext, type: Type): strin
   let addedCounter = 0;
   let removeCounter = 0;
   const retval: string[] = [];
-  for (const version of apiVersions) {
+  for (let i = 0; i < apiVersions.length; i++) {
+    const version = apiVersions[i];
     if (addedCounter < addedOnVersions.length && version === addedOnVersions[addedCounter]) {
       added = true;
       addedCounter++;
@@ -137,7 +140,17 @@ export function getAvailableApiVersions(context: TCGCContext, type: Type): strin
       added = false;
       removeCounter++;
     }
-    if (added) retval.push(version);
+    if (added) {
+      // only add version smaller than config
+      if (
+        context.apiVersion === undefined ||
+        context.apiVersion === "latest" ||
+        context.apiVersion === "all" ||
+        apiVersions.indexOf(context.apiVersion) >= i
+      ) {
+        retval.push(version);
+      }
+    }
   }
   return retval;
 }
@@ -259,6 +272,11 @@ export interface TCGCContext {
   __api_versions?: string[];
   knownScalars?: Record<string, SdkBuiltInKinds>;
   diagnostics: readonly Diagnostic[];
+  __subscriptionIdParameter?: SdkParameter;
+  __rawClients?: SdkClient[];
+  apiVersion?: string;
+  __service_projection?: Map<Namespace, [Namespace, ProjectedProgram | undefined]>;
+  originalProgram: Program;
 }
 
 export function createTCGCContext(program: Program): TCGCContext {
@@ -266,6 +284,7 @@ export function createTCGCContext(program: Program): TCGCContext {
     program,
     emitterName: "__TCGC_INTERNAL__",
     diagnostics: [],
+    originalProgram: program,
   };
 }
 
@@ -344,4 +363,12 @@ export function isMultipartFormData(
   operation?: Operation
 ): boolean {
   return isMultipartOperation(context, operation) && isOperationBodyType(context, type, operation);
+}
+
+export function isSubscriptionId(context: TCGCContext, parameter: { name: string }): boolean {
+  return Boolean(context.arm) && parameter.name === "subscriptionId";
+}
+
+export function onClient(context: TCGCContext, parameter: { name: string }): boolean {
+  return isSubscriptionId(context, parameter) || isApiVersion(context, parameter);
 }