+++ packages/typespec-client-generator-core/src/package.ts
@@ -59,6 +59,7 @@ import {
 import { createDiagnostic } from "./lib.js";
 import {
   getClientNamespaceString,
+  getCrossLanguageDefinitionId,
   getCrossLanguagePackageId,
   getDefaultApiVersion,
   getEffectivePayloadType,
@@ -163,15 +164,16 @@ function getSdkLroServiceMethod<
     getSdkBasicServiceMethod<TOptions, TServiceOperation>(context, operation)
   );
 
-  basicServiceMethod.response.type = diagnostics.pipe(
-    getClientTypeWithDiagnostics(context, metadata.logicalResult)
-  );
-  basicServiceMethod.response.resultPath =
-    metadata.logicalPath ??
-    (metadata.envelopeResult !== metadata.logicalResult &&
-    basicServiceMethod.operation.verb === "post"
-      ? "result"
-      : undefined);
+  if (metadata.finalResult === undefined || metadata.finalResult === "void") {
+    basicServiceMethod.response.type = undefined;
+  } else {
+    basicServiceMethod.response.type = diagnostics.pipe(
+      getClientTypeWithDiagnostics(context, metadata.finalResult)
+    );
+  }
+
+  basicServiceMethod.response.resultPath = metadata.finalResultPath;
+
   return diagnostics.wrap({
     ...basicServiceMethod,
     kind: "lro",
@@ -259,20 +261,6 @@ function getSdkBasicServiceMethod<
     }
   }
 
-  // if there's an api version parameter, we want to bubble it up to the client
-  // we don't want it on the method level, but we will keep it on the service operation level
-  const apiVersionParam = methodParameters.find((x) => x.isApiVersionParam);
-  if (apiVersionParam && context.__api_version_parameter === undefined) {
-    context.__api_version_parameter = {
-      ...apiVersionParam,
-      name: "apiVersion",
-      nameInClient: "apiVersion",
-      isGeneratedName: apiVersionParam.name !== "apiVersion",
-      onClient: true,
-      optional: false,
-      clientDefaultValue: context.__api_version_client_default_value,
-    };
-  }
   const serviceOperation = diagnostics.pipe(
     getSdkServiceOperation<TOptions, TServiceOperation>(context, operation, methodParameters)
   );
@@ -283,7 +271,7 @@ function getSdkBasicServiceMethod<
     kind: "basic",
     name,
     access: getAccess(context, operation),
-    parameters: methodParameters.filter((x) => !x.isApiVersionParam),
+    parameters: methodParameters,
     description: getDocHelper(context, operation).description,
     details: getDocHelper(context, operation).details,
     operation: serviceOperation,
@@ -299,6 +287,7 @@ function getSdkBasicServiceMethod<
     getResponseMapping: function getResponseMapping(): string | undefined {
       return undefined; // currently we only return a value for paging or lro
     },
+    crossLanguageDefintionId: getCrossLanguageDefinitionId({ ...operation, name }),
   });
 }
 
@@ -328,6 +317,9 @@ function getClientDefaultApiVersion<
   context: SdkContext<TOptions, TServiceOperation>,
   client: SdkClient | SdkOperationGroup
 ): string | undefined {
+  if (context.apiVersion && !["latest", "all"].includes(context.apiVersion)) {
+    return context.apiVersion;
+  }
   let defaultVersion = getDefaultApiVersion(context, client.service)?.value;
   if (!defaultVersion) {
     // eslint-disable-next-line deprecation/deprecation
@@ -353,6 +345,9 @@ function getSdkInitializationType<
   }
   if (context.__api_version_parameter) {
     properties.push(context.__api_version_parameter);
+  }
+  if (context.__subscriptionIdParameter) {
+    properties.push(context.__subscriptionIdParameter);
   }
   const namePrefix = client.kind === "SdkClient" ? client.name : client.groupPath;
   const name = `${namePrefix.split(".").at(-1)}Options`;
@@ -419,15 +414,17 @@ function getSdkMethods<TOptions extends object, TServiceOperation extends SdkSer
   for (const operationGroup of listOperationGroups(context, client)) {
     // We create a client accessor for each operation group
     const operationGroupClient = diagnostics.pipe(createSdkClientType(context, operationGroup));
+    const name = `get${operationGroup.type.name}`;
     retval.push({
       kind: "clientaccessor",
       parameters: [],
-      name: `get${operationGroup.type.name}`,
+      name,
       description: getDocHelper(context, operationGroup.type).description,
       details: getDocHelper(context, operationGroup.type).details,
       access: "internal",
       response: operationGroupClient,
       apiVersions: getAvailableApiVersions(context, operationGroup.type),
+      crossLanguageDefintionId: getCrossLanguageDefinitionId({ ...operationGroup.type, name }),
     });
   }
   return diagnostics.wrap(retval);